requirements.txt
aiohttp
aiosignal
asn1crypto
async-timeout
attrs
azure-common
azure-communication-email
azure-core
azure-mgmt-core
azure-storage-file-share
blinker
cachetools
certifi
cffi
chardet
charset-normalizer
click
cloud-sql-python-connector
cmake
contextlib2
cryptography
dataclasses-json
deptry
exceptiongroup
execnet
faiss-cpu
faiss-gpu
filelock
Flask
Flask-Login
Flask-SQLAlchemy
frozenlist
fsspec
google-auth
GPUtil
greenlet
gunicorn
html2text
huggingface-hub
idna
iniconfig
isodate
itsdangerous
Jinja2
joblib
langchain
langchainplus-sdk
lit
MarkupSafe
marshmallow
marshmallow-enum
mock
mpmath
msrest
multidict
mypy-extensions
networkx
nltk
numexpr
numpy
nvidia-cublas-cu11
nvidia-cuda-cupti-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-cudnn-cu11
nvidia-cufft-cu11
nvidia-curand-cu11
nvidia-cusolver-cu11
nvidia-cusparse-cu11
nvidia-nccl-cu11
nvidia-nvtx-cu11
oauthlib
openai
openapi-schema-pydantic
packaging
path
path.py
pathspec
pg8000
Pillow
pluggy
pyasn1
pyasn1-modules
pycparser
pydantic
pytest
pytest-shutil
python-dateutil
PyYAML
regex
requests
requests-oauthlib
rsa
safetensors
scikit-learn
scipy
scramp
sentence-transformers
sentencepiece
six
SQLAlchemy
sympy
tenacity
termcolor
threadpoolctl
tokenizers
tomli
torch
torchvision
tqdm
transformers
triton
typing-inspect
typing_extensions
urllib3
Werkzeug
xtarfile
yarl

